stream.segmented: move CompatThreadPoolExecutor
